HIGH_BYTE and LOW_BYTE cast to uint8_t now

+++ util.h
@@ -77,9 +77,9 @@
 
 // Get the high or low byte of a two byte value (in a very paranoid way :)
 #define HIGH_BYTE(two_byte_value) \
-  ((((uint16_t) two_byte_value) >> BITS_PER_BYTE) & ((uint16_t) 0x00ff))
+  ((uint8_t) ((((uint16_t) two_byte_value) >> BITS_PER_BYTE) & 0x00ff))
 #define LOW_BYTE(two_byte_value) \
-  (((uint16_t) two_byte_value) & ((uint16_t) 0x00ff))
+  ((uint8_t) (((uint16_t) two_byte_value) & 0x00ff))
 
 // Macros for more readable/writable binary bit patterns {{{1
 #define B0 0
